requirements/requirements.txt
Django==2.2
django-ckeditor==6.2.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.10.1
Markdown==3.3.6
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
